Improved (Error: "You need to first build the unit test app") (#15)
* Improved (Error: "You need to first build the unit test app")

If the program in the debug configuration is not found, 'build', 'switch' or 'ignore' are now options.

'switch': Switch to a different debug config (now a quick pick).

'build': Allows the user to pick a build task to run (quick pick). When the build task completes, the tests are reloaded.

'ignore': Subdues the error.

Platform support was also added. The platform specific 'program' field in the debug configuration takes presidence.

A nuance was added to recognise applications with native extensions when the extension is not provided in the debug configuration.

* Better feedback when selecting 1st debug config